pyproject.toml
[project]
name = "fragment-completion"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.8",
    "numpy>=2.4.2",
    "open-clip-torch>=3.2.0",
    "pillow>=12.1.1",
    "timm>=1.0.24",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    "transformers>=5.1.0",
    "scikit-learn>=1.5.0",
]
